requirements.txt
yt-dlp
flet
Pillow
pyinstaller
